add remove murren from group funk
