+++ murr_chat/routing.py
@@ -1,12 +1,12 @@
 from django.urls import path
 
 from murr_chat.consumers.group import GroupChatConsumer
-from .consumers.chat import ChatConsumer
+from .consumers.chat import GroupConsumer
 
 
 websocket_urls = [
 
     path('ws/chat/', GroupChatConsumer),
-    path('ws/chat/<str:group_id>/', ChatConsumer)
+    path('ws/chat/<str:group_id>/', GroupConsumer)
 
 ]
