+++ templates/murr_game/murr_game.html
@@ -96,7 +96,7 @@
                 },
                 come_to_tawern() {
 
-                    {# подключаться к сокету tawern #}
+                    {# подключаемся к сокету tawern #}
 
                     let group_id = '1';
                     let ws = new WebSocket('ws://127.0.0.1:8000/ws/chat/' + group_id + '/');
@@ -106,29 +106,41 @@
                     ws.onopen = () => {
                         console.log("Соединение с сокетом установлено.");
 
-                        let event_come_to_tawern = '{"event": "add.chat.member", "data": {"user_id": "' +
+                        let command = '{"event": "add.chat.member", "data": {"user_id": "' +
                             this.user_id +
                             '"}}';
+                        ws.send(command);
 
-                        ws.send(event_come_to_tawern);
 
-
-                        this.show_tawern = (this.show_tawern === false);
+                        this.show_tawern = true;
                         console.log("Добро пожаловать в таверну!");
                     };
 
                     ws.onmessage = (data) => {
                         let data_from_django = JSON.parse(data.data);
 
+                        {# Отлавливаем ответ с бекенда #}
                         if (data_from_django.event === "send.message") {
-                            let x = data_from_django.data.message.toString();
-                            this.logs.push(x);
+                            let new_message = data_from_django.data.message.toString();
+                            this.logs.unshift(new_message);
+                        }
+                        if (data_from_django.event === "murren.in.tawern.list") {
+                            console.log('murren.in.tawern.list');
+                            console.log(data_from_django);
+                            console.log('murren.in.tawern.list');
+                        }
+                        if (data_from_django.event === "leave.group") {
+
+                            this.logs.unshift('Вы вышли из группы');
+                            this.show_murren_group = false;
                         }
                     };
 
                     ws.onclose = () => {
                         console.log('Сокет закрыт')
-                    }
+                    };
+
+                    this.show_murren_group = true;
                 },
                 send_message() {
                     let message = '{"event": "send.message", "data": {"message": "' +
@@ -136,6 +148,21 @@
                         '"}}';
                     this._ws.send(message);
                     this.murren_message_input = '';
+
+                },
+
+
+                event_leave_group() {
+                    let command = '{"event": "leave.group", "data": {}}';
+                    this._ws.send(command);
+                    console.log('На бекенд отправлена задача' + command);
+                },
+
+                show_murren_in_tawern() {
+
+                    let command = '{"event": "murren.in.tawern.list", "data": {}}';
+                    this._ws.send(command);
+                    console.log('На бекенд отправлена задача' + command);
 
                 },
                 create_group_by_name() {
@@ -168,7 +195,7 @@
 
                     };
 
-                    this.show_murren_group = true;
+
                     this.show_create_group = false;
 
 
@@ -204,7 +231,7 @@
                     };
 
                 },
-                leave_group(){
+                leave_group() {
 
                     this._ws.close();
                     console.log('Вы покидаете группу');
@@ -240,9 +267,7 @@
                 show_create_group_menu() {
                     return this.show_create_group
                 },
-                show_murren_group_menu() {
-                    return this.show_murren_group
-                },
+
                 get_user_id() {
                     return this.user_id
                 },
